bug/module_average_incorrect (#33)

+++ app/src/main/java/mobile/joehonour/newcastleuniversity/unitracker/domain/calculations/IProvideModuleCalculations.kt
@@ -5,5 +5,5 @@ import mobile.joehonour.newcastleuniversity.unitracker.domain.models.Module
 interface IProvideModuleCalculations
 {
     fun calculatePercentageCompleteOf(module: Module) : Double
-    fun calculateCurrentAverageGradeOf(module: Module) : Double
+    fun calculateCurrentAverageGradeOf(module: Module) : Int
 }
